+++ frontend/src/models/ui/hotkeys-defined.ts
@@ -12,6 +12,13 @@ export const HK_EDIT_RANDOM: HotKeyShortcuts = [["N", HotKeyMask.SHIFT]];
 export const HK_EDIT_REPLACE: HotKeyShortcuts = [["R", HotKeyMask.SHIFT]];
 export const HK_PASTE: HotKeyShortcuts = [["V", HotKeyMask.CTRL], ["Paste"]];
 export const HK_REDO: HotKeyShortcuts = [["Y", HotKeyMask.CTRL], ["Redo"]];
+export const HK_TOOL_FLOOD: HotKeyShortcuts = [["F"]];
+export const HK_TOOL_LINE: HotKeyShortcuts = [["L"]];
+export const HK_TOOL_PEN: HotKeyShortcuts = [["1"]];
+export const HK_TOOL_PEN2: HotKeyShortcuts = [["2"]];
+export const HK_TOOL_PEN3: HotKeyShortcuts = [["3"]];
+export const HK_TOOL_RECT: HotKeyShortcuts = [["R"]];
+export const HK_TOOL_SELECTION: HotKeyShortcuts = [["S"]];
 export const HK_UNDO: HotKeyShortcuts = [["Z", HotKeyMask.CTRL], ["Undo"]];
 export const HK_ZOOM_IN: HotKeyShortcuts = [
   // Num pad <+>